Fix reset for moto.service backends
